Use uint64 for iterations
Resolves https://github.com/k6io/k6/pull/1863#discussion_r641465216

+++ lib/executor/ramping_arrival_rate.go
@@ -202,12 +202,12 @@ func (varr *RampingArrivalRate) Init(ctx context.Context) error {
 // this executor, taking into account the configured execution segment.
 // Unlike the local iteration number returned by getNextLocalIter(), this
 // iteration number will be unique across k6 instances.
-func (varr *RampingArrivalRate) getNextGlobalIter() int64 {
+func (varr *RampingArrivalRate) getNextGlobalIter() uint64 {
 	varr.iterMx.Lock()
 	defer varr.iterMx.Unlock()
 	varr.segIdx.Next()
 	// iterations are 0-based
-	return varr.segIdx.GetUnscaled() - 1
+	return uint64(varr.segIdx.GetUnscaled() - 1)
 }
 
 // cal calculates the  transtitions between stages and gives the next full value produced by the
